git-pimp: Code review or pull requests as patch email series
git-pimp generates a series of email messages from a range of commits,
with a cover letter summarizing the range.  Each commit email contains
everything necessary to reconstruct the commit in another repository
(provided it has all parents of the commit in question), plus extra
optional information such as diffstats.

Cover letters are generated with git-mantle(1), commit emails with
git-format-patch(1), and all is sent out using git-mailz(1).