git bc: --reroll-count is a post-1.8.1.4 addition
(1.8.3.1 is happy already).  that means parent (v0.5) is broken
on RHEL-6, SLES-11-SP3, openSUSE-12.3, etc.

the fix is a little invasive and the mechanism a bit ugly:
git-format-patch *never* gets --reroll-count, the subject tag
gets modified instead.  i should be able to get rid of this hack
in a few years. :]